refactor: extract TaskItem component
